API changed for 64 bit user system rights

+++ src/java/netxms-client-api/src/main/java/org/netxms/api/client/Session.java
@@ -198,7 +198,7 @@ public interface Session
 	 * @return System-wide rights of currently logged in user
 	 */
 
-	public abstract int getUserSystemRights();
+	public abstract long getUserSystemRights();
 
 	/**
 	 * @return the passwordExpired
